修复每分钟步数不显示的问题

+++ app/src/main/java/com/hzp/pedometer/service/step/StepManager.java
@@ -160,7 +160,7 @@ public class StepManager implements StepDetector.OnStepCountListener {
      * 获得每分钟的步行速率
      * @return 步数每分钟
      */
-    public double getStepPerMin() {
+    public int getStepPerMin() {
         return stepPerMin;
     }
